+++ docs/usage/changelog.rst
@@ -6,6 +6,8 @@ Changelog
 0.6.3 - TBD
   * Added ``effects/run``, ``effects/stop`` and ``effects/status`` commands
   * Added ``power_toggle`` command
+  * Updated lifx-photons-core
+  * Shutdown of the server should be a bit more graceful now
 
 0.6.2 - 27 February 2020
   * Updated lifx-photons-core